% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_comparison.R
\name{best_model}
\alias{best_model}
\title{Name of the best model according to a comparison table}
\usage{
best_model(table)
}
\arguments{
\item{table}{Object returned by \code{\link{aic_table}}.}
}
\value{
Character with the name (key) of the model with the lowest criterion.
}
\description{
Name of the best model according to a comparison table
}
\examples{
data(growth_data)
mm <- fit_multimodel(growth_data, "age", "size",
                      models = c("von_bertalanffy", "gompertz", "logistic"))
comp <- aic_table(mm)
best_model(comp)
}
