% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_models.R
\name{gompertz}
\alias{gompertz}
\title{Gompertz growth model}
\usage{
gompertz(t, Linf, K, t0)
}
\arguments{
\item{t}{Vector of ages.}

\item{Linf}{Asymptotic size.}

\item{K}{Relative instantaneous growth rate at the inflection point.}

\item{t0}{Age at the inflection point of the curve (size = Linf/e).}
}
\value{
Vector of predicted sizes.
}
\description{
L(t) = Linf * exp(-exp(-K * (t - t0)))
}
\examples{
gompertz(t = 1:10, Linf = 85, K = 0.6, t0 = 1)
}
