% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_fitting.R
\name{logLik.growth_fit}
\alias{logLik.growth_fit}
\title{Extract the log-likelihood from a growth_fit object}
\usage{
\method{logLik}{growth_fit}(object, ...)
}
\arguments{
\item{object}{Object of class \code{"growth_fit"}.}

\item{...}{Unused.}
}
\value{
An object of class \code{"logLik"}, with the number of
  estimated parameters attached as its \code{"df"} attribute (see
  \code{\link[stats]{logLik}}).
}
\description{
Extract the log-likelihood from a growth_fit object
}
\examples{
data(growth_data)
fit <- fit_growth(growth_data, "age", "size", model = "von_bertalanffy")
logLik(fit)
}
