% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_grouped.R
\name{logLik.growth_fit_grouped}
\alias{logLik.growth_fit_grouped}
\title{Extract the log-likelihood from a growth_fit_grouped object}
\usage{
\method{logLik}{growth_fit_grouped}(object, ...)
}
\arguments{
\item{object}{Object of class \code{"growth_fit_grouped"}.}

\item{...}{Unused.}
}
\value{
An object of class \code{"logLik"}, with the total number of
  estimated parameters (common + free x groups) as its \code{"df"} attribute.
}
\description{
Extract the log-likelihood from a growth_fit_grouped object
}
\examples{
data(growth_data)
fit <- fit_growth_grouped(growth_data, "age", "size", "group", free = "all")
logLik(fit)
}
