% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_fitting.R
\name{predict.growth_fit}
\alias{predict.growth_fit}
\title{Predict sizes from a growth_fit object}
\usage{
\method{predict}{growth_fit}(object, new_t = NULL, n = 200, ...)
}
\arguments{
\item{object}{Object of class \code{"growth_fit"}.}

\item{new_t}{Vector of ages (or times) at which to predict. If \code{NULL},
a regular grid over the observed range is used.}

\item{n}{Number of grid points if \code{new_t} is \code{NULL}.}

\item{...}{Unused.}
}
\value{
Data frame with columns \code{t} and \code{y_pred}.
}
\description{
Predict sizes from a growth_fit object
}
\examples{
data(growth_data)
fit <- fit_growth(growth_data, "age", "size", model = "von_bertalanffy")
predict(fit, new_t = c(1, 5, 10))
head(predict(fit))
}
